Adds TestSetup
